move TestGetContraction to test_helpers
also cleaned long lines in test_shapetracker and enabled the line length check
